pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "apscheduler>=3.11.0",
    "asyncpg>=0.30.0",
    "python-dotenv>=1.1.1",
    "python-telegram-bot==20.0",
]
